Extract reload-guards region into sections/reload-guards.js
Third region extraction of P-09. 60 LOC.

The pending-reload bookkeeping (_pendingReloadNodes, _pendingReloadInPlace) was mutated from three dashboard-main.js regions: the tab-switch path set pendingInPlace=true, refreshAll cleared both, reloadNode pushed nodes into the set. ESM live bindings don't permit importer-side reassignment of let/var exports, so the extraction replaces both with a single 'reloadState' object (same trick as P-07's termState). Three call sites updated. _reloadInPlace, reloadNode, _runnerViewers all picked up matching 'export' keywords to support the circular import.
